Fix escaped text breaking ampersand

+++ database/migrations/2024_07_31_215400_remove_unused_slug.php
@@ -30,6 +30,9 @@ return new class () extends Migration {
             'timelines',
         ];
         foreach ($tables as $tableName) {
+            if (!Schema::hasColumn($tableName, 'slug')) {
+                continue;
+            }
             Schema::table($tableName, function (Blueprint $table) {
                 $table->dropColumn('slug');
             });
